Move off Tcl proxy functions to separate file and make its syntax clearer
