+++ namd/colvars/src/Makefile.namd
@@ -22,6 +22,7 @@ COLVARSLIB = \
 	$(DSTDIR)/colvarparse.o \
 	$(DSTDIR)/colvarproxy.o \
 	$(DSTDIR)/colvarproxy_replicas.o \
+	$(DSTDIR)/colvarproxy_tcl.o \
 	$(DSTDIR)/colvarscript.o \
 	$(DSTDIR)/colvarscript_commands.o \
 	$(DSTDIR)/colvarscript_commands_bias.o \
