+++ vmd/src/colvars_files.pl
@@ -34,6 +34,7 @@ $colvars_defines = " -DVMDCOLVARS";
                     'colvarparse.C',
                     'colvarproxy.C',
                     'colvarproxy_replicas.C',
+                    'colvarproxy_tcl.C',
                     'colvarproxy_vmd.C',
                     'colvarscript.C',
                     'colvarscript_commands.C',
@@ -59,6 +60,7 @@ $colvars_defines = " -DVMDCOLVARS";
                     'colvarparams.h',
                     'colvarparse.h',
                     'colvarproxy.h',
+                    'colvarproxy_tcl.h',
                     'colvarproxy_vmd.h',
                     'colvarscript.h',
                     'colvarscript_commands.h',
